Now a specific battleMission is assigned to each icon on the 2d battleMap BUG: But there is a memory leak. When you load a lot of battles, a lot of memory is used. That is, the memory of the last battle is not released

+++ Defend It/PresentationLayer/Flows/TabbarFlow/Coordinator/TabbarCoordinatorOutput.swift
@@ -11,5 +11,5 @@ protocol TabbarCoordinatorOutput: AnyObject {
     var startMainFlow: (() -> Void)? { get set }
     
     
-    var startBattleFlow: ((Int) -> Void)? { get set }
+    var startBattleFlow: (() -> Void)? { get set }
 }
